requirements.txt
requests
pandas
matplotlib
python-dotenv
langchain
langchain-google-genai
langchain-mcp
langgraph
